pyproject.toml
[tool.poetry]
name = "course-work-7"
version = "0.1.0"
description = ""
authors = ["Alex Urapin <kulob1@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
openapi = "^2.0.0"
celery = "^5.4.0"
redis = "^5.2.1"
ruamel-yaml = "^0.18.8"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
